requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
environ==1.0
idna==3.10
Markdown==3.6
pillow==10.3.0
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.1
urllib3==2.3.0
whitenoise==6.7.0